Fix tests parsing credit amounts as JSON numbers after u128
The u128 change made the API serialize amounts as decimal strings, but several
tests still deserialized `balance` as i64, so `/status` JSON parsing panicked —
api_status_endpoint had been red in CI since the u128 commit (my earlier fix only
covered the job-bid string fields, and I failed to confirm CI went green).

- local_cluster api_status_endpoint: Status.balance String, parse to i128 for the
  assertion.
- ce-deploy e2e (ignored): same fix in both Status structs.
- ce-chain bench: signed_transfer amount -> u128 (cargo test doesn't compile benches,
  so this was latent; cargo bench now builds).

+++ crates/ce-chain/benches/chain.rs
@@ -9,7 +9,7 @@ fn tmp_identity(tag: &str) -> Identity {
     Identity::load_or_generate(&dir).unwrap()
 }
 
-fn signed_transfer(from: &Identity, to: ce_identity::NodeId, amount: u64) -> Tx {
+fn signed_transfer(from: &Identity, to: ce_identity::NodeId, amount: u128) -> Tx {
     let kind = TxKind::Transfer { from: from.node_id(), to, amount };
     let data = bincode::serialize(&kind).unwrap();
     let sig = from.sign(&data);
